+++ server/src/App/Models/Traits/WithPdf.php
@@ -29,11 +29,10 @@ trait WithPdf
 
         $i18n = new I18n(Config::getSettings('defaultLang'));
         $fileName = sprintf(
-            '%s-%s-%s-%s.pdf',
+            '%s-%s-%s.pdf',
             $i18n->translate($this->_modelName),
             slugify($company['name']),
-            $model->number,
-            slugify($model->Beneficiary->full_name)
+            $model->id
         );
         if (isTestMode()) {
             $fileName = sprintf('TEST-%s', $fileName);